Separated video components into there own files
