+++ smartHub_app/components/lists/FeaturesImageList.tsx
