+++ smartHub_app/components/pages/TakePhotoPage.tsx
@@ -1,252 +1,92 @@
 import React, {Component} from 'react';
-import {StyleSheet, View, Text, TouchableOpacity, Dimensions} from 'react-native';
+import {StyleSheet, View, Dimensions} from 'react-native';
 import {WebView} from 'react-native-webview'
 import axios from 'axios'
-import Toast from 'react-native-toast-message'
+import Toast, {BaseToast} from 'react-native-toast-message'
 import { getAddressString } from '../../utils/utilities';
-import getToastConfig from '../configurations/toastConfig';
+import Stream from '../video/Streaming';
 
 var width : number = Dimensions.get('window').width;
 
-export default class TakePhoto extends Component<{route: any, navigation: any}, {checkStream: boolean, profileId: number, deviceIP: String, phoneNumber: String,userEmail: String, profileName: String}>{
 
-    constructor(props: any){
+export default class TakePhoto extends Component<{ route: any, navigation: any }, { profileId: number, phoneNumber: String, deviceIP: String, userEmail: String, profileName: String}>{
+
+    constructor(props: any) {
         super(props);
-        this.state= ({
-            checkStream: false,
+        this.state = ({
             deviceIP: "",
-            profileId: 0,
             userEmail: "",
-            phoneNumber: "",
             profileName: "",
+            phoneNumber: "",
+            profileId: 0
         })
-    
-        this.beginStream = this.beginStream.bind(this);
-        this.stopStream = this.stopStream.bind(this);    
     }
-
-
     getDeviceInfo = async () => {
         //console.log(this.props.route);
         let collection: any = {}
         collection.device_id = this.props.route.params.device_id;
         await axios.post(getAddressString() + '/devices/getDeviceInfo', collection).then((response) => {
             this.setState({deviceIP: response.data.device[0].device_address,
-            userEmail: response.data.device[0].user_email,
-            profileName: response.data.device[0].profile_name,
-            profileId: response.data.device[0].profile_id,
-            phoneNumber: response.data.device[0].phone_number})
-            console.log(response.data)
+                userEmail: response.data.device[0].user_email,
+                profileName: response.data.device[0].profile_name,
+                profileId: response.data.device[0].profile_id,
+                phoneNumber: response.data.device[0].phone_number})
         }, (error) => {
             console.log(error);
         })
     }
-    
-    beginStream = () => {
-        if(!this.state.checkStream){
-            var url = 'http://' + this.state.deviceIP + ':4000/video/start_stream';
-            if(this.state.deviceIP !== 'lukessmarthub.ddns.net' &&  this.state.deviceIP !== "petepicam1234.zapto.org" && this.state.deviceIP !== "leohescamera.ddns.net"){
-                alert(this.props.route.params.device_name + ' not compatible for live streaming.')
-                return;
-            }
-            axios.post(url).then((response) => {
-                this.setState({checkStream: true});
-                Toast.show({
-                    type: 'success',
-                    text1: 'Processing Request Please Wait...',
-                    visibilityTime: 5000
-                })
-                setTimeout(() => {
-                    Toast.show({
-                        type: 'success',
-                        text1: 'The Stream Is Live!',
-                        text2: 'Enjoy!',
-                        visibilityTime: 4000
-                    })
-                }
-                ,
-                5000);
-            }, (error) => {
-                console.log(error);
-            })
-        }else if(this.state.checkStream){
-            Toast.show({
-                type: 'success',
-                text1: 'The Stream Is Already Live!',
-                text2: 'Click on the video player to view the stream.',
-                visibilityTime: 2000
-            })
-        }
-    }
-    
-    stopStream = () => {
-        if(this.state.checkStream){
-            var url = 'http://' + this.state.deviceIP + ':4000/video/stop_stream';
-            if(this.state.deviceIP !== 'lukessmarthub.ddns.net' &&  this.state.deviceIP !== "petepicam1234.zapto.org" && this.state.deviceIP !== "leohescamera.ddns.net"){
-                alert(this.props.route.params.device_name + ' not compatible for live streaming.')
-                return;
-            }
-            axios.post(url).then((response) => {
-                this.setState({checkStream: false})
-                Toast.show({
-                    type: 'error',
-                    text1: 'Stop Stream Clicked!',
-                    text2: 'The stream is no longer live.',
-                    visibilityTime: 2000
-                });
-            }, (error) => {
-                console.log(error);
-            })
-        }else{
-            Toast.show({
-                type: 'success',
-                text1: 'The Stream has already stopped!',
-                visibilityTime: 2000
-            })
-        }
-    }
-
-    // stopStreamOnBackClick = () => {
-    //     var url = 'http://' + this.state.deviceIP + ':4000/video/stop_stream';
-    //     if(this.state.deviceIP !== 'petepicam1234.zapto.org' && this.state.deviceIP !== "leohescamera.ddns.net"){
-    //         return;
-    //     }
-    //     if(this.state.hasStarted){
-    //         axios.post(url).then((response) => {
-    //             this.setState({hasStarted: false})
-    //         }, (error) => {
-    //             console.log(error);
-    //         })
-    //     }
-    // }
-
-    takePhoto = () => {
-        var url = 'http://' + this.state.deviceIP + ':4000/video/takeFaceImage';
-        if(this.state.deviceIP !== 'lukessmarthub.ddns.net' &&  this.state.deviceIP !== "petepicam1234.zapto.org" && this.state.deviceIP !== "leohescamera.ddns.net"){
-            alert(this.props.route.params.device_name + ' not compatible for photo taking.')
-            return;
-        }
-
-        if(!this.state.checkStream){
-            alert("You must begin the stream first");
-            return;
-        }
-
-        let collection: any = {}
-        collection.user_email = this.state.userEmail;
-        collection.profile_name = this.state.profileName;
-        collection.component_name = "Faces";
-        collection.profile_id = this.state.profileId;
-
-        Toast.show({
-            type: 'success',
-            text1:"Processing Please Wait...",
-            visibilityTime: 6000
-        });
-        axios.post(url, collection).then((response) => {
-            console.log(response.data)
-            Toast.show({
-                type: 'success',
-                text1: response.data,
-                visibilityTime: 2000
-            });
-        }, ({error,response}) : any => {
-            Toast.show({
-                type: 'success',
-                text1: response.data,
-                text2: "Please Try Again.",
-                visibilityTime: 3000
-            });
-        })
-    }
-
-
-    // checkStream = () => {
-    //     var url = 'http://' + this.state.deviceIP + ':4000/video/stream_check';
-    //     axios.post(url).then((response) => {
-    //         this.setState({checkStream: response.data.streaming})
-    //     }, (error) => {
-    //         console.log(error);
-    //     })
-    // }
 
     componentDidMount = async() => {
-        await this.getDeviceInfo();
-       // await this.checkStream();
+            await this.getDeviceInfo();
+             this.props.navigation.setOptions({
+                headerTitle: this.props.route.params.device_name,
+            })
     }
 
     render(){
+        const toastConfig = {
+            success: ({ text1, text2, ...rest } : any) => (
+              <BaseToast
+                {...rest}
+                style={{ borderLeftColor: '#FF9900', backgroundColor: "#fff" }}
+                contentContainerStyle={{ paddingHorizontal: 15 }}
+                text1Style={{
+                  fontSize: 18,
+                  fontWeight: 'bold'
+                }}
+                text2Style={{
+                    color: "#000",
+                    fontSize: 12
+                }}
+                text1={text1}
+                text2={text2}
+              />
+            ),
+
+            error: ({ text1, text2, ...rest } : any) => (
+                <BaseToast
+                  {...rest}
+                  style={{ borderLeftColor: '#FF9900', backgroundColor: "#fff" }}
+                  contentContainerStyle={{ paddingHorizontal: 15 }}
+                  text1Style={{
+                    fontSize: 18,
+                    fontWeight: 'bold'
+                  }}
+                  text2Style={{
+                      color: "#000",
+                      fontSize: 10
+                  }}
+                  text1={text1}
+                  text2={text2}
+                />
+              )
+        }
         return(
             <View style={{flex:1, backgroundColor: "#222222"}}>
-                <Toast style={{zIndex: 1}} config={getToastConfig()} ref={(ref) => Toast.setRef(ref)} />
-                <WebView
-                    style={{
-                        flex: 1,
-                    }}
-                    originWhitelist={['*']}
-                    source={{html: '<iframe style="box-sizing: border-box; width: 100%; height: 100%; border: 15px solid #FF9900; background-color: #222222"; src="http://' + this.state.deviceIP + ':4000/watch.html" frameborder="0" allow="autoplay encrypted-media" allowfullscreen></iframe>'}} 
-                    mediaPlaybackRequiresUserAction={false}
-                />
-                <View style={{flexDirection: 'row', justifyContent: 'center', alignItems: 'center', padding: 50, paddingBottom: 30}}>
-                    <TouchableOpacity
-                        style={styles.pillButton}
-                        onPress= {
-                            this.beginStream
-                        }>
-                        <Text style={{fontSize: 20}}>Begin Stream</Text>
-                    </TouchableOpacity>
-                    <TouchableOpacity
-                        style={styles.pillButton}
-                        onPress={
-                            this.stopStream
-                        }>
-                        <Text style={{fontSize: 20}}>Stop Stream</Text>
-                    </TouchableOpacity>
-                </View>
-                <View style={{alignItems: 'center'}}>
-                    <TouchableOpacity
-                        style={styles.photoButton}
-                        onPress={this.takePhoto}>
-                        <Text style={{fontSize: 20}}>Take Photo</Text>
-                    </TouchableOpacity>
-                </View>
-                <View style={{flexDirection: 'row', justifyContent: 'center', alignItems: 'center', paddingTop: 30, paddingBottom: 80}}>
-                </View>
+                <Toast style={{zIndex: 1}} config={toastConfig} ref={(ref) => Toast.setRef(ref)} />
+                <Stream type={3} deviceIP={this.state.deviceIP} deviceId={this.props.route.params.device_id} navigation={this.props.navigation} userEmail={this.state.userEmail} profileName={this.state.profileName}
+                profileId={this.state.profileId} phoneNumber={this.state.phoneNumber}/>                
             </View>
         );
     }
 }
-
-const styles = StyleSheet.create ({
-
-    pillButton: {
-        borderWidth:1,
-        justifyContent:'center',
-        alignItems:'center',
-        margin: 5,
-        width:175,
-        height:50,        
-        borderRadius:20,
-        backgroundColor: '#FF9900',
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 3 },
-        shadowOpacity: 0.5,
-        shadowRadius: 5, 
-    },
-
-    photoButton: {
-        borderWidth:1,
-        justifyContent:'center',
-        alignItems:'center',
-        margin: 5,
-        width:width-75,
-        height:50,        
-        borderRadius:20,
-        backgroundColor: '#FF9900',
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 3 },
-        shadowOpacity: 0.5,
-        shadowRadius: 5, 
-    },
-
-})
